+++ imgui-binding/src/main/java/imgui/ImFontGlyph.java
@@ -22,6 +22,7 @@ public final class ImFontGlyph implements ImDestroyable {
     }
 
     /*JNI
+        #include <stdint.h>
         #include <imgui.h>
         #include "jni_common.h"
 
@@ -37,7 +38,7 @@ public final class ImFontGlyph implements ImDestroyable {
 
     private native long nCreate(); /*
         ImFontConfig* imFontConfig = new ImFontConfig();
-        return (long)imFontConfig;
+        return (long)(intptr_t)imFontConfig;
     */
 
     private native void nDestroy(long ptr); /*
